// apps/backend/src/interfaces/user.interface.ts
export interface ProfilelUser {
    name: string;
    last_name: string;
    streak_days: number;
    experience_points: number;
};

export interface informationUser {
    name: string;
    last_name: string;
    document_type: string
    document: string;
    birthdate: Date;
    gender: string;
    email: string;
    phone_numer: string;
    country: string;
    department: string;
    municipality: string;
    neighborhood: string;
    commune: string;
    role: string;
    school: string;
    education_level: string;
    ethnic_group: string;
    conflict_victim: string;
}